+++ Whiteboard/Classes/SDK/WhiteSDK.m
@@ -28,7 +28,7 @@
 
 + (NSString *)version
 {
-    return @"2.16.41";
+    return @"2.16.42";
 }
 
 - (instancetype)initWithWhiteBoardView:(WhiteBoardView *)boardView config:(WhiteSdkConfiguration *)config commonCallbackDelegate:(nullable id<WhiteCommonCallbackDelegate>)callback audioMixerBridgeDelegate:( id<WhiteAudioMixerBridgeDelegate>)mixer